+++ libs/features/quest/src/quest-template-locale/quest-template-locale.component.ts
@@ -3,22 +3,19 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { QUEST_LOCALE, QuestTemplateLocale } from '@keira/shared/acore-world-model';
 import { MultiRowEditorComponent } from '@keira/shared/base-abstract-classes';
 import { EditorButtonsComponent, QueryOutputComponent, TopBarComponent } from '@keira/shared/base-editor-components';
-import { SingleValueSelectorBtnComponent } from '@keira/shared/selectors';
+import { GenericOptionSelectorComponent } from '@keira/shared/selectors';
 import { TranslateModule } from '@ngx-translate/core';
+import { NgxDatatableModule } from '@siemens/ngx-datatable';
 import { TooltipModule } from 'ngx-bootstrap/tooltip';
 import { QuestHandlerService } from '../quest-handler.service';
 import { QuestPreviewComponent } from '../quest-preview/quest-preview.component';
 import { QuestPreviewService } from '../quest-preview/quest-preview.service';
 import { QuestTemplateLocaleService } from './quest-template-locale.service';
-import { Model3DViewerComponent } from '@keira/shared/model-3d-viewer';
-import { AsyncPipe } from '@angular/common';
-import { NgxDatatableModule } from '@siemens/ngx-datatable';
 
 @Component({
   changeDetection: ChangeDetectionStrategy.OnPush,
   selector: 'keira-quest-template-locale',
   templateUrl: './quest-template-locale.component.html',
-  styleUrls: ['./quest-template-locale.component.scss'],
   standalone: true,
   imports: [
     TopBarComponent,
@@ -27,12 +24,10 @@ import { NgxDatatableModule } from '@siemens/ngx-datatable';
     FormsModule,
     ReactiveFormsModule,
     TooltipModule,
-    SingleValueSelectorBtnComponent,
     QuestPreviewComponent,
-    Model3DViewerComponent,
-    AsyncPipe,
     EditorButtonsComponent,
     NgxDatatableModule,
+    GenericOptionSelectorComponent,
   ],
 })
 export class QuestTemplateLocaleComponent extends MultiRowEditorComponent<QuestTemplateLocale> {
