+++ libs/shared/acore-world-model/src/index.ts
@@ -3,38 +3,38 @@ export * from './entities/area.type';
 export * from './entities/broadcast-text.type';
 export * from './entities/conditions.type';
 export * from './entities/creature-equip-template.type';
+export * from './entities/creature-formations.type';
 export * from './entities/creature-loot-template.type';
 export * from './entities/creature-onkill-reputation.type';
 export * from './entities/creature-questender.type';
 export * from './entities/creature-questitem.type';
 export * from './entities/creature-queststarter.type';
-export * from './entities/creature-spawn.type';
-export * from './entities/creature-formations.type';
 export * from './entities/creature-spawn-addon.type';
-export * from './entities/creature-template.type';
+export * from './entities/creature-spawn.type';
 export * from './entities/creature-template-addon.type';
 export * from './entities/creature-template-model.type';
 export * from './entities/creature-template-movement.type';
 export * from './entities/creature-template-resistance.type';
 export * from './entities/creature-template-spell.type';
+export * from './entities/creature-template.type';
 export * from './entities/creature-text.type';
 export * from './entities/disenchant-loot-template.type';
 export * from './entities/faction.type';
 export * from './entities/fishing-loot-template.type';
+export * from './entities/game-tele.type';
 export * from './entities/gameobject-loot-template.type';
 export * from './entities/gameobject-questender.type';
 export * from './entities/gameobject-questitem.type';
 export * from './entities/gameobject-queststarter.type';
-export * from './entities/gameobject-spawn.type';
 export * from './entities/gameobject-spawn-addon.type';
-export * from './entities/gameobject-template.type';
+export * from './entities/gameobject-spawn.type';
 export * from './entities/gameobject-template-addon.type';
-export * from './entities/game-tele.type';
-export * from './entities/gossip-menu.type';
+export * from './entities/gameobject-template.type';
 export * from './entities/gossip-menu-option.type';
+export * from './entities/gossip-menu.type';
 export * from './entities/holiday.type';
-export * from './entities/item-enchantment.type';
 export * from './entities/item-enchantment-template.type';
+export * from './entities/item-enchantment.type';
 export * from './entities/item-extended-cost.type';
 export * from './entities/item-limit-category.type';
 export * from './entities/item-loot-template.type';
@@ -53,17 +53,17 @@ export * from './entities/pickpocketing-loot-template.type';
 export * from './entities/prospecting-loot-template.type';
 export * from './entities/quest-offer-reward.type';
 export * from './entities/quest-request-items.type';
-export * from './entities/quest-template.type';
-export * from './entities/quest-template-locale.type';
 export * from './entities/quest-template-addon.type';
+export * from './entities/quest-template-locale.type';
+export * from './entities/quest-template.type';
 export * from './entities/reference-loot-template.type';
 export * from './entities/skill.type';
 export * from './entities/skinning-loot-template.type';
 export * from './entities/smart-scripts.type';
 export * from './entities/sound-entries.type';
-export * from './entities/spell.type';
 export * from './entities/spell-dbc.type';
 export * from './entities/spell-loot-template.type';
+export * from './entities/spell.type';
 
 export * from './flags/allowable-classes';
 export * from './flags/allowable-races';
@@ -120,17 +120,18 @@ export * from './options/item-class';
 export * from './options/item-honorrank';
 export * from './options/item-material';
 export * from './options/item-quality';
-export * from './options/text-type';
-export * from './options/text-range';
 export * from './options/item-sheath';
 export * from './options/movement-type';
 export * from './options/quest-info';
+export * from './options/quest-locale';
 export * from './options/quest-type';
 export * from './options/rank';
 export * from './options/spell-dbc-base-options';
 export * from './options/spell-dbc-effects-options';
 export * from './options/spell-mechanic';
 export * from './options/stat-type';
+export * from './options/text-range';
+export * from './options/text-type';
 export * from './options/totem-category';
 export * from './options/trainer-type';
 export * from './options/unit-class';
