+++ libs/shared/acore-world-model/src/options/quest-locale.ts
@@ -1,0 +1,11 @@
+import { Option } from '@keira/shared/constants';
+
+export const QUEST_LOCALE: Option[] = [
+  { value: 'deDE', name: 'German (Germany)' },
+  { value: 'esES', name: 'Spanish (Spain)' },
+  { value: 'esMX', name: 'Spanish (Mexico)' },
+  { value: 'frFR', name: 'French (France)' },
+  { value: 'ruRU', name: 'Russian (Russia)' },
+  { value: 'zhCN', name: 'Chinese (China)' },
+  { value: 'zhTW', name: 'Chinese (Taiwan)' },
+];
